+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Rename user status column header from Active to Status. Refs UIU-3610.
 * *BREAKING* Provide new `preferredContactTypeIds` multi-selection, require mobile phone when SMS is preferred. Bumps consumed `users` interface to 16.5. Refs UIU-3553.
 * Refactored src/settings/index.js to use Settings instead of custom SettingsPage. Refs UIU-3611.
+* Add localization for the renew error message. Refs UIU-3537.
 
 ## [13.0.2] (https://github.com/folio-org/ui-users/tree/v13.0.2) (2026-06-12)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v13.0.1...v13.0.2)